refactor: 도메인 컨트롤러 마다 뷰 따로 분리해서 뷰컨트롤러로 이동
