--- src/main/java/com/b1aboa/wedug/controller/HomeController.java
@@ -1,16 +1,0 @@
-package com.b1aboa.wedug.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping("/")
-public class HomeController {
-
-    @GetMapping
-    public String showIndexPage() {
-        return "index";
-    }
-
-}
